Update readme.txt
Changed the language used within 'Generating the IMG'. The language has been simplified and more specific with certain phrases such as ChromiumOS.img being shown to relate to something like 'Arnold the bat'. This ties the instructions to the previous 1.0 method for anyone coming from an alternative source/version and makes it easier to understand.