silence cp
currently the croissant script runs cp with -v on every invocation, causing every single file on the Chrome OS root to be printed.
this is an issue because the script is so verbose that even with a huge scrollback everything in a terminal will be pushed away, including the first lines of output of the script.
in addition, it's pointless anyway as cp will still display errors in silent mode.